+++ src/transit.js
@@ -8,7 +8,6 @@
 
 const Promise 			= require("bluebird");
 const _ 				= require("lodash");
-const { generateToken } = require("./utils");
 
 const P 				= require("./packets");
 const { Packet } 		= require("./packets");
@@ -1036,7 +1035,7 @@ class Transit {
 	 * @memberof Transit
 	 */
 	sendPing(nodeID, id) {
-		return this.publish(new Packet(P.PACKET_PING, nodeID, { time: Date.now(), id: id || generateToken() }))
+		return this.publish(new Packet(P.PACKET_PING, nodeID, { time: Date.now(), id: id || this.broker.generateUid() }))
 			.catch(/* istanbul ignore next */ err => this.logger.error(`Unable to send PING packet to '${nodeID}' node.`, err));
 	}